+++ common/client-core/src/client/topology_control/geo_aware_provider.rs
@@ -1,14 +1,14 @@
 use std::{collections::HashMap, fmt};
 
 use log::{debug, error, info};
-use nym_explorer_api_requests::{PrettyDetailedGatewayBond, PrettyDetailedMixNodeBond};
+use nym_explorer_client::{ExplorerClient, PrettyDetailedMixNodeBond};
 use nym_network_defaults::var_names::EXPLORER_API;
 use nym_topology::{
     nym_topology_from_detailed,
     provider_trait::{async_trait, TopologyProvider},
     NymTopology,
 };
-use nym_validator_client::client::{IdentityKey, MixId};
+use nym_validator_client::client::MixId;
 use rand::{prelude::SliceRandom, thread_rng};
 use serde::{Deserialize, Serialize};
 use tap::TapOptional;
@@ -18,55 +18,24 @@ use crate::config::GroupBy;
 
 const MIN_NODES_PER_LAYER: usize = 1;
 
-#[cfg(target_arch = "wasm32")]
-fn reqwest_client() -> Option<reqwest::Client> {
-    reqwest::Client::builder().build().ok()
-}
+fn create_explorer_client() -> Option<ExplorerClient> {
+    let Ok(explorer_api_url) = std::env::var(EXPLORER_API) else {
+        error!("Missing EXPLORER_API");
+        return None;
+    };
 
-#[cfg(not(target_arch = "wasm32"))]
-fn reqwest_client() -> Option<reqwest::Client> {
-    reqwest::Client::builder()
-        .timeout(std::time::Duration::from_secs(5))
-        .build()
-        .ok()
-}
+    let Ok(explorer_api_url) = explorer_api_url.parse() else {
+        error!("Failed to parse EXPLORER_API");
+        return None;
+    };
 
-// TODO: create a explorer-api-client
-async fn fetch_mixnodes_from_explorer_api() -> Option<Vec<PrettyDetailedMixNodeBond>> {
-    let explorer_api_url = std::env::var(EXPLORER_API).ok()?;
-    let explorer_api_url = Url::parse(&explorer_api_url)
-        .ok()?
-        .join("v1/mix-nodes")
-        .ok()?;
+    log::debug!("Using explorer-api url: {}", explorer_api_url);
+    let Ok(client) = nym_explorer_client::ExplorerClient::new(explorer_api_url) else {
+        error!("Failed to create explorer-api client");
+        return None;
+    };
 
-    debug!("Fetching: {}", explorer_api_url);
-    reqwest_client()?
-        .get(explorer_api_url)
-        .send()
-        .await
-        .ok()?
-        .json::<Vec<PrettyDetailedMixNodeBond>>()
-        .await
-        .ok()
-}
-
-// TODO: create a explorer-api-client
-async fn fetch_gateways_from_explorer_api() -> Option<Vec<PrettyDetailedGatewayBond>> {
-    let explorer_api_url = std::env::var(EXPLORER_API).ok()?;
-    let explorer_api_url = Url::parse(&explorer_api_url)
-        .ok()?
-        .join("v1/gateways")
-        .ok()?;
-
-    debug!("Fetching: {}", explorer_api_url);
-    reqwest_client()?
-        .get(explorer_api_url)
-        .send()
-        .await
-        .ok()?
-        .json::<Vec<PrettyDetailedGatewayBond>>()
-        .await
-        .ok()
+    Some(client)
 }
 
 #[derive(Copy, Clone, Hash, PartialEq, Eq, Serialize, Deserialize, Debug)]
@@ -83,6 +52,8 @@ pub enum CountryGroup {
 impl CountryGroup {
     // We map contry codes into group, which initially are continent codes to a first approximation,
     // but we do it manually to reserve the right to tweak this distribution for our purposes.
+    // NOTE: I did this quickly and it's not a complete list of all countries, but only those that
+    // were present in the network at the time. Please add more as needed.
     fn new(country_code: &str) -> Self {
         let country_code = country_code.to_uppercase();
         use CountryGroup::*;
@@ -235,23 +206,6 @@ fn group_mixnodes_by_country_code(
         })
 }
 
-fn group_gateways_by_country_code(
-    gateways: Vec<PrettyDetailedGatewayBond>,
-) -> HashMap<CountryGroup, Vec<IdentityKey>> {
-    gateways.into_iter().fold(
-        HashMap::<CountryGroup, Vec<IdentityKey>>::new(),
-        |mut acc, g| {
-            if let Some(ref location) = g.location {
-                let country_code = location.two_letter_iso_country_code.clone();
-                let group_code = CountryGroup::new(country_code.as_str());
-                let gateways = acc.entry(group_code).or_insert_with(Vec::new);
-                gateways.push(g.gateway.identity_key)
-            }
-            acc
-        },
-    )
-}
-
 fn log_mixnode_distribution(mixnodes: &HashMap<CountryGroup, Vec<MixId>>) {
     let mixnode_distribution = mixnodes
         .iter()
@@ -259,15 +213,6 @@ fn log_mixnode_distribution(mixnodes: &HashMap<CountryGroup, Vec<MixId>>) {
         .collect::<Vec<_>>()
         .join(", ");
     debug!("Mixnode distribution - {}", mixnode_distribution);
-}
-
-fn log_gateway_distribution(gateways: &HashMap<CountryGroup, Vec<IdentityKey>>) {
-    let gateway_distribution = gateways
-        .iter()
-        .map(|(k, v)| format!("{}: {}", k, v.len()))
-        .collect::<Vec<_>>()
-        .join(", ");
-    debug!("Gateway distribution - {}", gateway_distribution);
 }
 
 fn check_layer_integrity(topology: NymTopology) -> Result<(), ()> {
@@ -303,7 +248,7 @@ impl GeoAwareTopologyProvider {
         filter_on: GroupBy,
     ) -> GeoAwareTopologyProvider {
         log::info!(
-            "Creating geo-aware topology provider with filter on {:?}",
+            "Creating geo-aware topology provider with filter on {}",
             filter_on
         );
         nym_api_urls.shuffle(&mut thread_rng());
@@ -337,13 +282,14 @@ impl GeoAwareTopologyProvider {
         // Also fetch mixnodes cached by explorer-api, with the purpose of getting their
         // geolocation.
         debug!("Fetching mixnodes from explorer-api...");
-        let Some(mixnodes_from_explorer_api) = fetch_mixnodes_from_explorer_api().await else {
+        let explorer_client = create_explorer_client()?;
+        let Ok(mixnodes_from_explorer_api) = explorer_client.get_mixnodes().await else {
             error!("failed to get mixnodes from explorer-api");
             return None;
         };
 
         debug!("Fetching gateways from explorer-api...");
-        let Some(gateways_from_explorer_api) = fetch_gateways_from_explorer_api().await else {
+        let Ok(gateways_from_explorer_api) = explorer_client.get_gateways().await else {
             error!("failed to get mixnodes from explorer-api");
             return None;
         };
@@ -382,27 +328,14 @@ impl GeoAwareTopologyProvider {
         let mixnode_distribution = group_mixnodes_by_country_code(mixnodes_from_explorer_api);
         log_mixnode_distribution(&mixnode_distribution);
 
-        let gateway_distribution = group_gateways_by_country_code(gateways_from_explorer_api);
-        log_gateway_distribution(&gateway_distribution);
-
         let Some(filtered_mixnode_ids) = mixnode_distribution.get(&filter_on) else {
             error!("no mixnodes found for: {}", filter_on);
-            return None;
-        };
-
-        let Some(filtered_gateway_ids) = gateway_distribution.get(&filter_on) else {
-            error!("no gateways found for: {}", filter_on);
             return None;
         };
 
         let mixnodes = mixnodes
             .into_iter()
             .filter(|m| filtered_mixnode_ids.contains(&m.mix_id()))
-            .collect::<Vec<_>>();
-
-        let gateways = gateways
-            .into_iter()
-            .filter(|g| filtered_gateway_ids.contains(g.identity()))
             .collect::<Vec<_>>();
 
         let topology = nym_topology_from_detailed(mixnodes, gateways)