+++ nym-connect/native/android/app/src/main/java/net/nymtech/nyms5/ProxyWorker.kt
@@ -17,6 +17,10 @@ import androidx.work.WorkerParameters
 import androidx.work.workDataOf
 import com.github.kittinunf.fuel.Fuel
 import io.sentry.Sentry
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
@@ -32,7 +36,9 @@ class ProxyWorker(
     companion object Work {
         const val name = "nymS5ProxyWorker"
         const val workTag = "nymProxy"
-        val workId: UUID = UUID.randomUUID()
+        // it is very important to use a static UUID in order to allow WorkManager
+        // handling the proxy work as a unique work
+        val workId: UUID = UUID.fromString("cc785aa4-5775-4bf0-b870-39645e35e660")
 
         const val State = "State"
 
@@ -44,6 +50,8 @@ class ProxyWorker(
     }
 
     private val tag = "proxyWorker"
+
+    private val pingRate = 1000L
 
     private val spUrl = context.getString(R.string.sp_url)
 
@@ -135,7 +143,26 @@ class ProxyWorker(
                 Log.w(tag, "using a default service provider $defaultSp")
             }
 
-            nymProxy.start(serviceProvider ?: defaultSp, onStartCb, onStopCb)
+            withContext(Dispatchers.IO) {
+                val pingJob = launch {
+                    // this job will get automatically killed by the WorkManager once
+                    // the job has been terminated, so it's safe to use `while (true)`
+                    while (true) {
+                        nymProxy.ping()
+                        delay(pingRate)
+                    }
+                }
+
+                val proxyJob = launch {
+                    nymProxy.start(serviceProvider ?: defaultSp, onStartCb, onStopCb)
+                }
+
+                // wait for the underlying call to `startClient` to be released which means
+                // the connection has been terminated (`startClient` is a blocking call)
+                proxyJob.join()
+                // stop pinging
+                pingJob.cancel()
+            }
 
             // the state should be already set to DISCONNECTED at this point
             // but for the sake of it, reset it