+++ service-providers/network-requester/src/config/persistence.rs
@@ -7,6 +7,7 @@ use std::path::{Path, PathBuf};
 
 pub const DEFAULT_ALLOWED_LIST_FILENAME: &str = "allowed.list";
 pub const DEFAULT_UNKNOWN_LIST_FILENAME: &str = "unknown.list";
+pub const DEFAULT_DESCRIPTION_FILENAME: &str = "description.toml";
 
 #[derive(Debug, Deserialize, PartialEq, Eq, Serialize, Clone)]
 pub struct NetworkRequesterPaths {
@@ -18,6 +19,11 @@ pub struct NetworkRequesterPaths {
 
     /// Location of the file containing our unknown.list
     pub unknown_list_location: PathBuf,
+
+    /// Location of the file containing our description
+    // For upgrade use default if missing. On next config upgrade iteration, remove the serde(default)
+    #[serde(default)]
+    pub nr_description: PathBuf,
 }
 
 impl NetworkRequesterPaths {
@@ -28,6 +34,7 @@ impl NetworkRequesterPaths {
             common_paths: CommonClientPaths::new_default(base_dir),
             allowed_list_location: base_dir.join(DEFAULT_ALLOWED_LIST_FILENAME),
             unknown_list_location: base_dir.join(DEFAULT_UNKNOWN_LIST_FILENAME),
+            nr_description: base_dir.join(DEFAULT_DESCRIPTION_FILENAME),
         }
     }
 }