+++ service-providers/network-requester/src/config/template.rs
@@ -75,6 +75,9 @@ allowed_list_location = '{{ storage_paths.allowed_list_location }}'
 # Location of the file containing our unknown.list
 unknown_list_location = '{{ storage_paths.unknown_list_location }}'
 
+# Path to file containing description of this network-requester.
+nr_description = '{{ storage_paths.nr_description }}'
+
 
 [network_requester]